Add files via upload
This report is about the mediaeval memorability task, 
which seems to be the part of a medieval benchmarking initiative 
for evaluation. Participating members are meant to implement 
a system utilizing machine learning algorithms that would 
naturally anticipate memorability scores for videos in terms of 
features which emulates the probability of the videos being 
remembered. In relations to existing work within image memory 
probability, whereas image memorability was calculated right 
after being memorized, however, the dataset is composed of 
short and long-term memorability annotations which are 
utilized to calculate are efficient the system is.